fix: apply min-height and increase by content #61
refactor: extract subcomponents #61
