+++ src/Components/Common/Dialog/subcomponents/DialogBody/DialogBody.style.tsx
@@ -1,0 +1,12 @@
+import styled from '@emotion/styled';
+
+export const DialogBodyContainer = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  min-height: 3rem;
+  width: 100%;
+  padding: 0 1rem;
+`;
+
+export const DialogParagraph = styled.p``;
