+++ src/Components/Common/Dialog/subcomponents/DialogBody/DialogBody.tsx
@@ -1,0 +1,13 @@
+import { DialogBodyContainer, DialogParagraph } from './DialogBody.style';
+
+type DialogBodyProps = {
+  paragraph: string;
+};
+
+export function DialogBody({ paragraph }: DialogBodyProps) {
+  return (
+    <DialogBodyContainer>
+      <DialogParagraph>{paragraph}</DialogParagraph>
+    </DialogBodyContainer>
+  );
+}
