+++ src/Components/Common/Dialog/subcomponents/DialogBody/index.ts
@@ -1,0 +1,1 @@
+export * from './DialogBody';
