+++ src/Components/Common/Dialog/subcomponents/DialogFooter/DialogFooter.style.tsx
@@ -1,0 +1,10 @@
+import styled from '@emotion/styled';
+
+export const DialogFooterContainer = styled.div`
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  width: 100%;
+  padding: 0.75rem;
+  gap: 1rem;
+`;
