+++ src/Components/Common/Dialog/subcomponents/DialogFooter/index.ts
@@ -1,0 +1,1 @@
+export * from './DialogFooter';
