+++ src/Components/Common/Dialog/subcomponents/DialogHeader/DialogHeader.tsx
@@ -1,0 +1,36 @@
+import theme from '@/styles/theme';
+import {
+  DialogClose,
+  DialogHeaderContainer,
+  DialogHeaderTitle,
+} from './DialogHeader.style';
+
+type DialogHeaderProps = {
+  closeCB: () => void;
+  title: string;
+};
+
+export function DialogHeader({ title, closeCB }: DialogHeaderProps) {
+  return (
+    <DialogHeaderContainer>
+      <DialogHeaderTitle>{title}</DialogHeaderTitle>
+      <DialogClose onClick={closeCB}>
+        <svg
+          xmlns="http://www.w3.org/2000/svg"
+          width="24"
+          height="24"
+          viewBox="0 0 24 24"
+          strokeWidth="2"
+          stroke={theme.colors.gray700}
+          fill="none"
+          strokeLinecap="round"
+          strokeLinejoin="round"
+        >
+          <path stroke="none" d="M0 0h24v24H0z" fill="none"></path>
+          <path d="M18 6l-12 12"></path>
+          <path d="M6 6l12 12"></path>
+        </svg>
+      </DialogClose>
+    </DialogHeaderContainer>
+  );
+}
