+++ src/Components/Common/Dialog/subcomponents/DialogHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './DialogHeader';
