+++ src/Components/Common/Dialog/subcomponents/index.ts
@@ -1,0 +1,2 @@
+export * from './DialogFooter';
+export * from './DialogBody';
